# Modify attributes replacement for performance gains