--- a/R/categorize.R
+++ b/R/categorize.R
@@ -319,6 +319,8 @@ categorize.grouped_df <- function(x,
   # works only for dplyr >= 0.8.0
   grps <- attr(x, "groups", exact = TRUE)[[".rows"]]
 
+  attr_data <- attributes(x)
+
   # evaluate arguments
   select <- .select_nse(select,
     x,
@@ -353,7 +355,7 @@ categorize.grouped_df <- function(x,
     )
   }
   # set back class, so data frame still works with dplyr
-  x <- .replace_attrs(x, attributes(x))
+  x <- .replace_attrs(x, attr_data)
   x
 }