--- a/tests/testthat/test-data_reorder.R
+++ b/tests/testthat/test-data_reorder.R
@@ -23,5 +23,6 @@ test_that("data_reorder preserves attributes", {
   out2 <- data_reorder(out, 4:6)
   a2 <- attributes(out2)
 
-  expect_equal(names(a1), names(a2))
+  # attributes may not be in the same order
+  expect_true(all(names(a1) %in% names(a2)) && length(a1) == length(a2))
 })