@IndrajeetPatil 2022-10-12
Thanks, Etienne!

Nothing makes me happier than _O_(1) algorithms 😬